test: reframe full-clear accounting fixture away from #973 reproduction
